【AC】[AtCoder Beginner Contest 068]D - Decrease (Contestant ver.)

+++ abc068/arc079_b.cpp
@@ -30,7 +30,7 @@ const int mod = 1000000007;
 
 
 int main() {
-    int k;
+    ll k;
     cin >> k;
 
     int n = 50;
